Nest dashboard site_prism pages into dashboard folder
